refactor(menuable): use activatable
